+++ src/room/track/TrackPublication.ts
@@ -1,6 +1,6 @@
 import EventEmitter from 'eventemitter3';
 import log from '../../logger';
-import type { TrackInfo } from '../../proto/livekit_models';
+import type { SubscriptionError, TrackInfo } from '../../proto/livekit_models';
 import type { UpdateSubscription, UpdateTrackSettings } from '../../proto/livekit_rtc';
 import { TrackEvent } from '../events';
 import LocalAudioTrack from './LocalAudioTrack';
@@ -144,4 +144,5 @@ export type PublicationEventCallbacks = {
     status: TrackPublication.SubscriptionStatus,
     prevStatus: TrackPublication.SubscriptionStatus,
   ) => void;
+  subscriptionFailed: (error: SubscriptionError) => void;
 };